Split commands between GlobalCommands and NeovimEditorCommands
