+++ browser/src/Editor/NeovimEditor/NeovimEditor.tsx
@@ -30,7 +30,6 @@ import { PluginManager } from "./../../Plugins/PluginManager"
 
 import { IColors } from "./../../Services/Colors"
 import { commandManager } from "./../../Services/CommandManager"
-import { registerBuiltInCommands } from "./../../Services/Commands"
 import { Completion } from "./../../Services/Completion"
 import { Configuration, IConfigurationValues } from "./../../Services/Configuration"
 import { IDiagnosticsDataSource } from "./../../Services/Diagnostics"
@@ -179,13 +178,12 @@ export class NeovimEditor extends Editor implements IEditor {
         // Services
         const errorService = new Errors(this._neovimInstance)
 
-        registerBuiltInCommands(commandManager, this._neovimInstance)
-
         this._commands = new NeovimEditorCommands(
             commandManager,
             this._contextMenuManager,
             this._definition,
             this._languageIntegration,
+            this._neovimInstance,
             this._rename,
             this._symbols,
         )
