+++ browser/src/Services/Commands/index.ts
@@ -1,0 +1,1 @@
+export * from "./GlobalCommands"
